requirements.txt
pandas
PyMuPDF
python-docx
spacy 